configurations management task 3

+++ 0x0A-configuration_management/2-execute_a_command.pp
@@ -2,4 +2,5 @@
 
 exec { 'terminate_killmenow':
   command => 'pkill killmenow'
+  onlyif  => 'ps aux | grep killmenow'
 }
